Extract common methods to KafkaOffsetReaderBase
